Require core_ext/hash from ActiveSupport instead
When upgrading i18n in link handling, we get errors
coming from cequel. That's because it's requiring
a dependency that does not exist anymore in i18n > 1.18:

`i18n/core_ext/hash`.

After changing this line and using this commit's ref in link handling,
we can update i18n's version. I've tested this in staging by sending an email,
and verifying the open and click rates got recorded as expected.

As @cainwatson mentioned, this means we're good to go because "since this
gem is used for anything related to querying Cassandra if there were issues they would
 appear quickly and loudly."

I will also ship this in a canary for link handling.